backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2026.1.4
Django==4.2.28
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
gunicorn==23.0.0
packaging==26.0
pymongo==4.10.1
sqlparse==0.5.5
typing_extensions==4.13.2
whitenoise==6.7.0
